+++ noodles-sam/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+  * sam/record: Implement `std::error::Error::source` for errors.
+
   * sam/record/data/field: Add mutable getter for value (`Field::value_mut`).
 
 ## 0.22.1 - 2022-11-29